+++ src/main/java/web/Calculations.java
@@ -4,11 +4,11 @@ package web;
 		
 		static double a;
 		static double b;
-		static int bank;
+		static double bank;
 		static int period;
 		static double result;
 
-		Calculations (double a, double b, int bank, int period){
+		Calculations (double a, double b, double bank, int period){
 			
 			this.a = a;
 			this.b = b;
